Delete unnecessary files and update analytics package
