--- analytics/preference_queries.py
@@ -1,5 +1,0 @@
-import graphene
-from ..types.preference_types import Preference
-
-class PreferenceQuery(graphene.ObjectType):
-    getAllPreferences = graphene.List(Preference)
